Add multi-version validation for LLMInferenceService
Move validators from controller package to API packages following
kubebuilder multi-version best practices and existing kserve patterns.

Each API version now has its own validator that works directly with its own types, with separate webhook paths:
- v1alpha1: /validate-serving-kserve-io-v1alpha1-llminferenceservice
- v1alpha2: /validate-serving-kserve-io-v1alpha2-llminferenceservice

Changes:
- Create v1alpha1 validators in pkg/apis/serving/v1alpha1/
- Create v1alpha2 validators in pkg/apis/serving/v1alpha2/
- Add LLMInferenceServiceGVK to v1alpha1 package
- Add ParallelismSpec helper methods to v1alpha1
- Update main.go to register both version validators
- Update webhook manifests for versioned webhooks
- Remove duplicate validation packages from controller directory
