+++ config/crd/full/llmisvc/serving.kserve.io_llminferenceservices.yaml
@@ -57,6 +57,10 @@ spec:
                 model:
                   properties:
                     criticality:
+                      enum:
+                        - Critical
+                        - Standard
+                        - Sheddable
                       type: string
                     lora:
                       properties:
