+++ config/webhook/llmisvc/manifests.yaml
@@ -8,9 +8,9 @@ webhooks:
       service:
         name: $(webhookServiceName)
         namespace: $(kserveNamespace)
-        path: /validate-serving-kserve-io-v1alpha2-llminferenceservice
+        path: /validate-serving-kserve-io-v1alpha1-llminferenceservice
     failurePolicy: Fail
-    name: llminferenceservice.kserve-webhook-server.validator
+    name: llminferenceservice.kserve-webhook-server.v1alpha1.validator
     sideEffects: None
     admissionReviewVersions: [ "v1", "v1beta1" ]
     rules:
@@ -18,6 +18,24 @@ webhooks:
           - serving.kserve.io
         apiVersions:
           - v1alpha1
+        operations:
+          - CREATE
+          - UPDATE
+        resources:
+          - llminferenceservices
+  - clientConfig:
+      service:
+        name: $(webhookServiceName)
+        namespace: $(kserveNamespace)
+        path: /validate-serving-kserve-io-v1alpha2-llminferenceservice
+    failurePolicy: Fail
+    name: llminferenceservice.kserve-webhook-server.v1alpha2.validator
+    sideEffects: None
+    admissionReviewVersions: [ "v1", "v1beta1" ]
+    rules:
+      - apiGroups:
+          - serving.kserve.io
+        apiVersions:
           - v1alpha2
         operations:
           - CREATE
@@ -35,9 +53,9 @@ webhooks:
       service:
         name: $(webhookServiceName)
         namespace: $(kserveNamespace)
-        path: /validate-serving-kserve-io-v1alpha2-llminferenceserviceconfig
+        path: /validate-serving-kserve-io-v1alpha1-llminferenceserviceconfig
     failurePolicy: Fail
-    name: llminferenceserviceconfig.kserve-webhook-server.validator
+    name: llminferenceserviceconfig.kserve-webhook-server.v1alpha1.validator
     sideEffects: None
     admissionReviewVersions: [ "v1", "v1beta1" ]
     rules:
@@ -45,6 +63,25 @@ webhooks:
           - serving.kserve.io
         apiVersions:
           - v1alpha1
+        operations:
+          - CREATE
+          - UPDATE
+          - DELETE
+        resources:
+          - llminferenceserviceconfigs
+  - clientConfig:
+      service:
+        name: $(webhookServiceName)
+        namespace: $(kserveNamespace)
+        path: /validate-serving-kserve-io-v1alpha2-llminferenceserviceconfig
+    failurePolicy: Fail
+    name: llminferenceserviceconfig.kserve-webhook-server.v1alpha2.validator
+    sideEffects: None
+    admissionReviewVersions: [ "v1", "v1beta1" ]
+    rules:
+      - apiGroups:
+          - serving.kserve.io
+        apiVersions:
           - v1alpha2
         operations:
           - CREATE
