+++ pkg/apis/serving/v1alpha1/llm_inference_service_types_func.go
@@ -1,0 +1,59 @@
+/*
+Copyright 2025 The KServe Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package v1alpha1
+
+import (
+	"k8s.io/utils/ptr"
+)
+
+func (p *ParallelismSpec) IsPipelineParallel() bool {
+	if p == nil {
+		return false
+	}
+	return ptr.Deref(p.Pipeline, 0) > 0
+}
+
+func (p *ParallelismSpec) IsDataParallel() bool {
+	if p == nil {
+		return false
+	}
+	return ptr.Deref(p.Data, 0) > 0 || ptr.Deref(p.DataLocal, 0) > 0
+}
+
+func (p *ParallelismSpec) IsTensorParallel() bool {
+	if p == nil {
+		return false
+	}
+	return ptr.Deref(p.Tensor, 0) > 0
+}
+
+func (p *ParallelismSpec) GetSize() *int32 {
+	if p == nil {
+		return nil
+	}
+	if p.IsDataParallel() {
+		return ptr.To(max(
+			// p.Data / p.DataLocal
+			max(ptr.Deref(p.Data, 1), 1)/max(ptr.Deref(p.DataLocal, 1), 1),
+			1,
+		))
+	}
+	if p.IsPipelineParallel() {
+		return p.Pipeline
+	}
+	return nil
+}
