+++ pkg/apis/serving/v1alpha1/v1alpha1.go
@@ -39,6 +39,12 @@ var (
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme
 	SchemeBuilder = &scheme.Builder{GroupVersion: SchemeGroupVersion}
 
+	// LLMInferenceServiceGVK is the GroupVersionKind for LLMInferenceService
+	LLMInferenceServiceGVK = SchemeGroupVersion.WithKind("LLMInferenceService")
+
+	// LLMInferenceServiceConfigGVK is the GroupVersionKind for LLMInferenceServiceConfig
+	LLMInferenceServiceConfigGVK = SchemeGroupVersion.WithKind("LLMInferenceServiceConfig")
+
 	// AddToScheme is required by pkg/client/...
 	AddToScheme = SchemeBuilder.AddToScheme
 )
