--- pkg/controller/v1alpha2/llmisvc/validation/webhook_suite_test.go
@@ -1,39 +1,0 @@
-/*
-Copyright 2025 The KServe Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package validation_test
-
-import (
-	"testing"
-
-	"github.com/kserve/kserve/pkg/controller/v1alpha2/llmisvc/fixture"
-
-	pkgtest "github.com/kserve/kserve/pkg/testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestWebhooks(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "LLMInferenceService Controller Suite")
-}
-
-var envTest *pkgtest.Client
-
-var _ = SynchronizedBeforeSuite(func() {
-	envTest = fixture.SetupTestEnv()
-}, func() {})
