+++ test/webhooks/manifests.yaml
@@ -13,12 +13,33 @@ webhooks:
         path: /validate-serving-kserve-io-v1alpha1-llminferenceservice
     failurePolicy: Fail
     sideEffects: None
-    name: llminferenceservice.kserve-webhook-server.validator
+    name: llminferenceservice.kserve-webhook-server.v1alpha1.validator
     rules:
       - apiGroups:
           - serving.kserve.io
         apiVersions:
           - v1alpha1
+        operations:
+          - CREATE
+          - UPDATE
+        resources:
+          - llminferenceservices
+  - admissionReviewVersions:
+      - v1
+      - v1beta1
+    clientConfig:
+      service:
+        name: kserve-webhook-server-service
+        namespace: kserve
+        path: /validate-serving-kserve-io-v1alpha2-llminferenceservice
+    failurePolicy: Fail
+    sideEffects: None
+    name: llminferenceservice.kserve-webhook-server.v1alpha2.validator
+    rules:
+      - apiGroups:
+          - serving.kserve.io
+        apiVersions:
+          - v1alpha2
         operations:
           - CREATE
           - UPDATE
@@ -40,12 +61,34 @@ webhooks:
         path: /validate-serving-kserve-io-v1alpha1-llminferenceserviceconfig
     failurePolicy: Fail
     sideEffects: None
-    name: llminferenceserviceconfigs.kserve-webhook-server.validator
+    name: llminferenceserviceconfig.kserve-webhook-server.v1alpha1.validator
     rules:
       - apiGroups:
           - serving.kserve.io
         apiVersions:
           - v1alpha1
+        operations:
+          - CREATE
+          - UPDATE
+          - DELETE
+        resources:
+          - llminferenceserviceconfigs
+  - admissionReviewVersions:
+      - v1
+      - v1beta1
+    clientConfig:
+      service:
+        name: kserve-webhook-server-service
+        namespace: kserve
+        path: /validate-serving-kserve-io-v1alpha2-llminferenceserviceconfig
+    failurePolicy: Fail
+    sideEffects: None
+    name: llminferenceserviceconfig.kserve-webhook-server.v1alpha2.validator
+    rules:
+      - apiGroups:
+          - serving.kserve.io
+        apiVersions:
+          - v1alpha2
         operations:
           - CREATE
           - UPDATE
